requirements.txt
altair==6.0.0
annotated-types==0.7.0
asn1crypto==1.5.1
asttokens==3.0.0
atpublic==5.1
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.0.0
et_xmlfile==2.0.0
executing==2.2.0
frozendict==2.4.6
h11==0.14.0
icecream==2.1.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
multitasking==0.0.11
narwhals==1.30.0
numpy==2.2.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psycopg2-binary==2.9.11
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
scikit-learn
scipy
sec-edgar==0.0.2
selenium==4.31.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake==1.0.5
snowflake-connector-python==3.14.0
snowflake-snowpark-python==1.29.0
snowflake._legacy==1.0.0
snowflake.core==1.0.5
sortedcontainers==2.4.0
soupsieve==2.6
streamlit==1.54.0
tenacity==9.0.0
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==16.0
wsproto==1.2.0
yfinance==1.1.0
